Collapse long series description
This makes the collapsible description from video pages
reusable for series blocks and other components, with a
few context dependent adjustments. The reasoning for this
is that the description is now always shown inside the grey
box, where a long description distracts from the video. It
also let's us keep the ordering/layout option menus at the
top of the box without having to much empty space between
the menus and videos. We considered moving the menus below
the description but I think that results in an even more
uneven/unbalanced look and feel.
