+++ frontend/src/i18n/locales/en.yaml
@@ -143,7 +143,7 @@ video:
   starts: Starts
   starts-in: Starts {{duration}}
   description:
-    show-more: Show more
+    show-more: ...show more
     show-less: Show less
   embed:
     copy-embed-code-to-clipboard: Copy embed code to clipboard
